feat(Example NextJS): Add navigation and footer components, refactor page data fetching to use dotCMSClient
