+++ examples/nextjs/src/app/[[...slug]]/page.js
@@ -1,12 +1,12 @@
 import { cache } from "react";
 
 import { Page } from "@/pages/Page";
-import { getPage } from "@/utils/getPage";
+import { getDotCMSPage } from "@/utils/getDotCMSPage";
 
 export async function generateMetadata({ params, searchParams }) {
     try {
         const path = params?.slug?.join("/") || "/";
-        const { pageAsset } = await getPage(path, searchParams);
+        const { pageAsset } = await getDotCMSPage(path, searchParams);
         const page = pageAsset.page;
         const title = page?.friendlyName || page?.title;
 
@@ -22,6 +22,6 @@ export async function generateMetadata({ params, searchParams }) {
 
 export default async function Home({ params, searchParams }) {
     const path = params?.slug?.join("/") || "/";
-    const pageContent = await getPage(path, searchParams);
+    const pageContent = await getDotCMSPage(path, searchParams);
     return <Page pageContent={pageContent} />;
 }
