+++ examples/nextjs/src/components/footer/components/blogs.js
