+++ examples/nextjs/src/components/footer/components/destinations.js
