+++ examples/nextjs/src/components/footer/footer.js
