+++ examples/nextjs/src/components/navigation.js
