+++ examples/nextjs/src/utils/dotCMSClient.js
@@ -1,6 +1,6 @@
 import { createDotCMSClient } from "@dotcms/client/next";
 
-export const dotClient = createDotCMSClient({
+export const dotCMSClient = createDotCMSClient({
     dotcmsUrl: process.env.NEXT_PUBLIC_DOTCMS_HOST,
     authToken: process.env.NEXT_PUBLIC_DOTCMS_AUTH_TOKEN,
     siteId: "59bb8831-6706-4589-9ca0-ff74016e02b2",
