+++ examples/nextjs/src/utils/queries.js
@@ -1,0 +1,31 @@
+export const blogQuery = `
+    search(query: "+contenttype:Blog +live:true", limit: 3) {
+        title
+        identifier
+        ... on Blog {
+            inode
+            image {
+            fileName
+            }
+            urlMap
+            modDate
+            urlTitle
+        }
+    }
+`;
+
+export const destinationQuery = `
+    search(query: "+contenttype:Destination +live:true", limit: 3) {
+        title
+        identifier
+            ... on Destination {
+                inode
+                image {
+                fileName
+                }
+                urlMap
+                modDate
+                url
+        }
+    }
+`;
